(bzr r3008.10.3)

+++ unity-shared/PluginAdapter.cpp
@@ -1436,7 +1436,7 @@ void PluginAdapter::AddProperties(GVariantBuilder* builder)
   unity::variant::BuilderWrapper wrapper(builder);
   wrapper.add(GetScreenGeometry())
          .add("workspace_count", WorkspaceCount())
-         .add("active_window", GetActiveWindow())
+         .add("active_window", (guint64)GetActiveWindow())
          .add("screen_grabbed", IsScreenGrabbed())
          .add("scale_active", IsScaleActive())
          .add("scale_active_for_group", IsScaleActiveForGroup())
